-Implemented tests for calculateGrassPercentage and its stub

+++ src/main/java/edu/cps2002/mazegame/map/HazardousMap.java
@@ -2,7 +2,7 @@ package edu.cps2002.mazegame.map;
 
 public class HazardousMap extends Map{
     @Override
-    boolean setWaterPercentage(float waterPercentage) {
+    boolean setWaterPercentage(double waterPercentage) {
         //if percentage is not between 25% and 35%
         if(waterPercentage < 25 || waterPercentage > 35){
             return false;
