+++ src/main/java/edu/cps2002/mazegame/map/SafeMap.java
@@ -2,7 +2,7 @@ package edu.cps2002.mazegame.map;
 
 public class SafeMap extends Map{
     @Override
-    boolean setWaterPercentage(float waterPercentage) {
+    boolean setWaterPercentage(double waterPercentage) {
         //if percentage is greater than 10% or less than 0%
         if(waterPercentage < 0 || waterPercentage > 10){
             return false;
